REFACTOR: Extract a `Toolbar` component.
