+++ packages/frontend/src/App.tsx
@@ -12,7 +12,7 @@ import { Match, Switch, createResource, lazy, useContext } from "solid-js";
 import type { AppRouter } from "backend/src/index.js";
 import { RPCContext, RepoContext } from "./api";
 import { newModelDocument } from "./document/types";
-import { helpPage } from "./help";
+import { helpPage } from "./page/help_page";
 import { TheoryLibraryContext, stdTheories } from "./stdlib";
 
 const serverUrl: string = import.meta.env.VITE_BACKEND_HOST;
@@ -104,11 +104,11 @@ const routes: RouteDefinition[] = [
     },
     {
         path: "/help",
-        component: helpPage(lazy(() => import("./help/pages/index.mdx"))),
+        component: helpPage(lazy(() => import("./help/index.mdx"))),
     },
     {
         path: "/help/credits",
-        component: helpPage(lazy(() => import("./help/pages/credits.mdx"))),
+        component: helpPage(lazy(() => import("./help/credits.mdx"))),
     },
 ];
