+++ packages/frontend/src/help/credits.mdx
