--- packages/frontend/src/help/index.ts
@@ -1,2 +1,0 @@
-export * from "./help_button";
-export * from "./help_page";
