+++ packages/frontend/src/index.css
@@ -75,18 +75,3 @@ h3 {
 .side-panel {
     background: #f8f8f8;
 }
-
-.toolbar {
-    display: flex;
-    flex-direction: row;
-    gap: 5px;
-    padding: 5px;
-
-    & .filler {
-        flex: 1;
-    }
-
-    & .lucide {
-        color: dimgray;
-    }
-}
