+++ packages/frontend/src/page/toolbar.css
@@ -1,0 +1,14 @@
+.toolbar {
+    display: flex;
+    flex-direction: row;
+    gap: 5px;
+    padding: 5px;
+
+    & .filler {
+        flex: 1;
+    }
+
+    & .lucide {
+        color: dimgray;
+    }
+}
